reorganize file structure
This repo originally contained just the frontend mockup code, but now that I'm working on the race engine I compartmentalized into frontend and engine folders. In the future I'm assuming we'll use this repo for the rest of the project so these first efforts are going into engine/prototype and frontend/mockup folders preemptively.